fix:detail page hook
